requirements.txt
# LLM Frameworks
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# LlamaIndex
llama-index>=0.11.0
llama-index-readers-file>=0.4.0
llama-index-llms-openai>=0.3.0
llama-index-llms-openai-like>=0.2.0
llama-index-embeddings-openai>=0.3.0

# Optional local embedding dependencies live in requirements.local-embedding.txt

# PDF reading (for resume)
pypdf>=4.0.0

# Web
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12

# Storage
aiosqlite>=0.20.0

# Speech-to-text (Qiniu OSS + DashScope cloud ASR)
qiniu>=7.12.0

# Auth
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# Web search (Copilot Company Researcher)
tavily-python>=0.3.0

# Copilot 实时 ASR (DashScope qwen3-asr-flash-realtime)
websockets>=12.0

# Tencent Cloud VPR (可选，用于候选人声纹识别)
tencentcloud-sdk-python-common>=3.0.1000

# VAD 切段（C 库，非神经网络）
webrtcvad>=2.0.10

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
rich>=13.0.0
